added pagination of products and implemented delete user feature and added dynamic reviews

+++ public/js/admin.js
@@ -2,17 +2,18 @@ const deleteProduct = (btn) => {
     const prodId = btn.parentNode.querySelector('[name=id]').value;
     console.log(prodId);
     const elem = btn.parentNode.parentNode;
-    console.log(elem);
-    fetch(`/admin/product/${prodId}` , {
-        method : "DELETE",
-    })
-    .then(result => {
-        return result.json();
-    })
-    .then(data => {
-        elem.remove();
-    })
-    .catch(err => {
-        console.log(err);
-    })
+    if(window.confirm("are you sure ?")){
+        fetch(`/admin/product/${prodId}` , {
+            method : "DELETE",
+        })
+        .then(result => {
+            return result.json();
+        })
+        .then(data => {
+            elem.remove();
+        })
+        .catch(err => {
+            console.log(err);
+        })
+    }
 };
